feat(hard link): move IsDir() to a separate file
